packages/mcp-video-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-video-server"
version = "0.1.0"
requires-python = ">=3.11"
description = "MCP server for video analysis — frame extraction, motion detection, and more"
dependencies = [
    "mcp>=1.0.0",
    "opencv-python-headless>=4.10.0",
    "numpy>=1.24.0",
    "Pillow>=11.0.0",
]

[project.optional-dependencies]
transcription = ["groq>=0.4.0"]
whisper = ["openai-whisper>=20231117"]
charts = ["matplotlib>=3.8.0"]

[project.scripts]
mcp-video-server = "mcp_video_server.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_video_server"]
